% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.r
\name{createAnnotationGrid}
\alias{createAnnotationGrid}
\title{createAnnotationGrid}
\usage{
createAnnotationGrid(cases, vals)
}
\description{
This funciton is meant to create a grid with an annotation

to be display concurrently with another type of graphic
}
\seealso{
\code{\link{waterfallPlot}}
}
