% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_survival.r
\name{forestPlot}
\alias{forestPlot}
\title{Create a forest plot of hazard ratios}
\usage{
forestPlot(vars, HRs, lowerCI, upperCI)
}
\arguments{
\item{vars}{Names of variables}

\item{lowerCI}{Lower CI estiate for each HR}

\item{upperCI}{Upper CI estimate for each HR}

\item{HR}{Hazard ratio point estimate for each variable (HR)}
}
\description{
Create a forest plot of hazard ratios
}
