% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.r
\name{gghist}
\alias{gghist}
\title{GGPlot histogram of variable x, with some basic statistics on graph}
\usage{
gghist(x, tit = "", bw = NA, displayStats = FALSE)
}
\arguments{
\item{x}{Variable to create histogram from}

\item{tit}{Optional title of graph}
}
\description{
GGPlot histogram of variable x, with some basic statistics on graph
}
\seealso{
\code{\link{histwStats}}
}
