% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_survival.r
\name{ggkm}
\alias{ggkm}
\title{GGPlot based Survival Curves by a Variable. Report p-values by log-rank & Cox Model}
\usage{
ggkm(event_time, event, max, var, lgnd, tit = "Survival")
}
\arguments{
\item{event_time}{Time to Event}

\item{event}{Event occured or not (0 or 1)}

\item{max}{Maximum time to display on graph}

\item{var}{Variable to split Survival Curves by (should be categorical or ordinal)}

\item{lgnd}{Legend text (optional)}

\item{tit}{Title of graph  (default = "Survival")}
}
\value{
p-value from log-rank test
}
\description{
GGPlot based Survival Curves by a Variable. Report p-values by log-rank & Cox Model
}
\seealso{
\code{\link{qkm}}, \code{\link{qkm2}}, \code{\link{qkm_var_med}}, \code{\link{qkm_var_anal}}
}
