% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{group_var}
\alias{group_var}
\title{Take a continous variable, and cut it into multiple points, by cut points}
\usage{
group_var(xvar, cp)
}
\arguments{
\item{xvar}{Variable to divide}

\item{cp:}{Cut points to divide variable by, firs group is < cp[1], 2nd group is >=cp[1], <cp[2], etc.}
}
\value{
ordinal scale 0,1,2,3,4....
}
