% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.r
\name{histwStats}
\alias{histwStats}
\title{Displays histogram of variable x, with some basic statistics on graph}
\usage{
histwStats(x, tit = "", xtit = "")
}
\arguments{
\item{x}{Variable to create histogram from}

\item{tit}{Optional title of graph}
}
\description{
Displays histogram of variable x, with some basic statistics on graph
}
\seealso{
\code{\link{gghist}}
}
