% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{list2matrix}
\alias{list2matrix}
\title{Take a list of vectors and convert it to a matrix}
\usage{
list2matrix(lst)
}
\arguments{
\item{lst}{Variable that is a list of vectors}
}
\value{
a matrix with same data
}
\description{
Take a list of vectors and convert it to a matrix
}
