% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.r
\name{myggboxplot}
\alias{myggboxplot}
\title{Create a box plot from a data frame}
\usage{
myggboxplot(dat, groupCol, groups, valCol, groupLabels, tit = "",
  colCol = "", violin = 0)
}
\arguments{
\item{dat}{Data matrix (contains Group Column; Val Column)}

\item{groupCol}{Name of column in dat, that contains information on group id}

\item{valCol}{Column with continous values to compare between two groups}

\item{groupLabels}{Labels to use for graphing}

\item{tit}{Title for graph (default "")}

\item{colCol}{Column to use for coloring data points (optinal)}

\item{violen}{display violen plot instead of boxplots}
}
\value{
ggplot graph
}
\description{
Create a box plot from a data frame
}
\seealso{
\code{\link{tTestandGraph}}, \code{\link{myswarmplot}}
}
