% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.r
\name{myswarmplot}
\alias{myswarmplot}
\title{Create a swarm plot to compare different groups}
\usage{
myswarmplot(dat, groupCol, groups, valCol, groupLabels, tit = "",
  colCol = "")
}
\arguments{
\item{dat}{Data matrix (contains Group Column; Val Column)}

\item{groupCol}{Name of column in dat, that contains information on group id}

\item{valCol}{Column with continous values to compare between two groups}

\item{groupLabels}{Labels to use for graphing}

\item{tit}{Title for graph (default "")}

\item{colCol}{Column to use for coloring data points (optinal)}
}
\value{
ggplot graph
}
\description{
Create a swarm plot to compare different groups
}
\seealso{
\code{\link{myggboxplot}},  \code{\link{tTestandGraph}}
}
