% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.r
\name{mytextplot}
\alias{mytextplot}
\title{Create a traditional R graphic object that just displays text}
\usage{
mytextplot(txt, size = 1)
}
\arguments{
\item{txt}{Text to display in graphical object}

\item{size}{Size of text}
}
\description{
Create a traditional R graphic object that just displays text
}
