% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_survival.r
\name{qkm}
\alias{qkm}
\title{Create a Kaplan Meier Survival Curve, print table of results}
\usage{
qkm(event_time, event, max = 5, tit = "Survival", xlabv = "Years",
  NNT = FALSE)
}
\arguments{
\item{event_time}{Time to Event}

\item{event}{Event occured or not (0 or 1)}

\item{max}{Maximum time to display on graph (default = 5)}

\item{tit}{Title of graph (default 'Suvival')}

\item{xlabv}{X-axis label (default: 'Years')}

\item{NNT}{Display number at risk (boolean)}
}
\value{
Model form survfit
}
\description{
Create a Kaplan Meier Survival Curve, print table of results
}
\seealso{
\code{\link{qkm2}}, \code{\link{ggkm}}, \code{\link{qkm_var_med}}, \code{\link{qkm_var_anal}}
}
