% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_survival.r
\name{qkm2}
\alias{qkm2}
\title{Survival Curves by a Variable. displays p-values by log-rank & cox model}
\usage{
qkm2(event_time, event, max, var, lgnd, tit = "Survival")
}
\arguments{
\item{event_time}{Time to Event}

\item{event}{Event occured or not (0 or 1)}

\item{lgnd}{Legend text (optional)}

\item{tit}{Title of graph  (default = 'Survival')}

\item{Max}{Maximum time to display on graph}

\item{Var}{Variable to split Survival Curves by (should be categorical or ordinal)}
}
\value{
P-value from log-rank test
}
\description{
Survival Curves by a Variable. displays p-values by log-rank & cox model
}
\seealso{
\code{\link{qkm}}, \code{\link{ggkm}}, \code{\link{qkm_var_med}}, \code{\link{qkm_var_anal}}
}
