% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_survival.r
\name{qkm_var_anal}
\alias{qkm_var_anal}
\title{Display 4 panels with analysis of a quantiative variable and survival}
\usage{
qkm_var_anal(event_time, event, max, var, tit)
}
\arguments{
\item{event_time}{Time to Event}

\item{event}{Event occured or not (0 or 1)}

\item{max}{Maximum time to display on graph}

\item{var}{Variable to split Survival Curves by (should be continous)}

\item{tit}{title of graph  (optional)}
}
\value{
P-value from log-rank test
}
\description{
Display 4 panels with analysis of a quantiative variable and survival
}
\seealso{
\code{\link{qkm}}, \code{\link{qkm2}}, \code{\link{ggkm}}, \code{\link{qkm_var_med}}
}
