% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_survival.r
\name{qkm_var_med}
\alias{qkm_var_med}
\title{Display Survival curve of quantitiative variable split @ median}
\usage{
qkm_var_med(event_time, event, max, var, tit)
}
\arguments{
\item{event_time}{Time to Event}

\item{event}{Event occured or not (0 or 1)}

\item{max}{Maximum time to display on graph}

\item{var}{Variable to split Survival Curves by (should be continous)}

\item{tit}{title of graph  (optional)}
}
\value{
P-value from log-rank test
}
\description{
Display Survival curve of quantitiative variable split @ median
}
\seealso{
\code{\link{qkm}}, \code{\link{qkm2}}, \code{\link{ggkm}}, \code{\link{qkm_var_anal}}
}
