% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_survival.r
\name{qkm_var_quant}
\alias{qkm_var_quant}
\title{Display Survival curve of quantitiative variable split by quantiles}
\usage{
qkm_var_quant(event_time, event, max, var, tit)
}
\arguments{
\item{event}{Event occured or not (0 or 1)}

\item{max}{Maximum time to display on graph}

\item{var}{Variable to split Survival Curves by (should be continous)}

\item{tit}{title of graph  (optional)}

\item{event_time:}{Time to Event}
}
\value{
P-value from log-rank test
}
\description{
Display Survival curve of quantitiative variable split by quantiles
}
\seealso{
\code{\link{qkm}}, \code{\link{qkm2}}, \code{\link{ggkm}}, \code{\link{qkm_var_anal}}
}
