% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.r
\name{qscat}
\alias{qscat}
\title{Quick scatter plot with linear regression line, and annotation at bottom of graph}
\usage{
qscat(x, y, xax = "x", yax = "y", main = "", bw = 0, displayStat = 1,
  dodraw = TRUE)
}
\description{
Quick scatter plot with linear regression line, and annotation at bottom of graph
}
