% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.r
\name{tTestandGraph}
\alias{tTestandGraph}
\title{Test for differences in a variable between two groups}
\usage{
tTestandGraph(dat, groupCol, groups, valCol, groupLabels, tit = "",
  colCol = "", swarm = FALSE, devoff = TRUE, dodraw = TRUE)
}
\arguments{
\item{dat}{Data matrix (contains Group Column; Val Column)}

\item{groupCol}{Name of column in dat, that contains information on group id}

\item{valCol}{Column with continous values to compare between two groups}

\item{groupLabels}{Labels to use for graphing}

\item{tit}{Title for graph (default "")}

\item{colCol}{Column to use for coloring data points (optinal)}

\item{swarm}{display indiivudal points}

\item{devoff}{TBD}

\item{dodraw}{TBD}
}
\value{
conglomerate graph object (from arrangeGrob)
}
\description{
Compares two groups for differences in a variable using a T-test and Wilcox test

Display corresponding graph with multiple display options
}
\seealso{
\code{\link{myggboxplot}},  \code{\link{myswarmplot}}
}
