% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{vexcel}
\alias{vexcel}
\title{write data frame to temporary csv and quickly open in excel}
\usage{
vexcel(dat)
}
\arguments{
\item{dat}{Data frame to write to a temporary csv to open up in excel}
}
\description{
write data frame to temporary csv and quickly open in excel
}
