% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.r
\name{waterfallPlot}
\alias{waterfallPlot}
\title{Create a waterfall plot for cases, using values}
\usage{
waterfallPlot(cases, vals, dec = FALSE)
}
\arguments{
\item{cases}{List of cases for waterfall plot}

\item{vals}{values to use in waterfall plot}

\item{dec}{sort cases in decreasing order}
}
\description{
Create a waterfall plot for cases, using values
}
\seealso{
\code{\link{createAnnotationGrid}}
}
