% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileio.r
\name{write_cox_to_csv}
\alias{write_cox_to_csv}
\title{Output results of a cox model to Excel (CSV file)}
\usage{
write_cox_to_csv(filename, model)
}
\arguments{
\item{filename}{file to save results to}

\item{model}{Cox model from coxph}
}
\description{
Output results of a cox model to Excel (CSV file)
}
\seealso{
\code{\link{write_uva_to_csv}}
}
